requirements.txt
langgraph
langchain-openai
python-dotenv
requests
beautifulsoup4
